fix(api): sync the metering test doubles; finish the models_models deletion
Two root causes behind the seven red lanes on the first CI run:

1. test_llm_metering patched calculate_token_cost with an AsyncMock after the
   function went sync, so the un-awaited coroutine became the "cost", the
   except path booked $0, and the baseline failed — taking the test shard and
   all three diff-scoped mutation shards down with it. The doubles are plain
   mocks now. (test_accounting's doubles were converted in the previous
   commit; this file was the one I missed in the sweep.)

2. Deleting the catalog left ModelProvider with zero users anywhere, which the
   strict vulture gate rightly flagged. Gone — and with it models_models.py
   held nothing but DevModelOption, which moves beside LLMProviderName in
   agents/llm/types.py (the module it already imported from); the file is
   deleted and its three importers repointed.

Local: 12129 unit + fallback-chain tests pass, vulture gate clean, mypy clean.
