+++ apps/api/app/agents/llm/types.py
@@ -2,6 +2,7 @@
 
 from collections.abc import Callable
 from enum import StrEnum
+from typing import Any
 
 from langchain_core.language_models import LanguageModelInput
 from langchain_core.messages import AIMessage
@@ -16,6 +17,25 @@ class LLMProviderName(StrEnum):
     GEMINI = "gemini"
     OPENROUTER = "openrouter"
     CUSTOM = "custom"
+
+
+class DevModelOption(TypedDict):
+    """One entry of the DEV-ONLY model menu (``constants.llm.DEV_MODEL_OPTIONS``).
+
+    A TypedDict, not a model: it is a fixed in-process shape that is only ever
+    spread onto a LangGraph configurable, so it crosses no validation boundary.
+
+    ``model_kwargs`` and ``reasoning``'s effort payload stay ``dict[str, Any]``
+    because that is exactly how ``ChatOpenRouter`` declares the fields they are
+    bound to — free-form OpenRouter request params, not a shape we own.
+    """
+
+    #: Keyed the same as ``PROVIDER_MODELS``/``PROVIDER_PRIORITY`` — the enum is
+    #: what stops the menu naming a lane the client cannot resolve.
+    provider: LLMProviderName
+    model: str
+    model_kwargs: dict[str, Any] | None
+    reasoning: bool
 
 
 class LLMProviderKey(StrEnum):
